add manage button to edit page
