+++ src/products/components/ProductVariantChannels/ChannelsAvailabilityCard/AvailabilityCard.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { useIntl } from "react-intl";
 
 import { variantDetailsChannelsAvailabilityCardMessages as messages } from "./../messages";
+import { Channel, ProductChannelListing } from "./../types";
 import { ChannelsList } from "./ChannelsList";
 import { ChannelsListItem } from "./ChannelsListItem";
-import { Channel, ProductChannelListing } from "./types";
 import CardContainer from "./VariantDetailsChannelsAvailabilityCardContainer";
 
 interface AvailabilityCardProps {
