--- src/products/components/ProductVariantChannels/ChannelsAvailabilityCard/types.ts
@@ -1,10 +1,0 @@
-import {
-  ProductVariantCreateDataQuery,
-  ProductVariantFragment,
-} from "@saleor/graphql";
-
-export type Product = ProductVariantCreateDataQuery["product"];
-export type Variant = ProductVariantFragment;
-export type ProductChannelListing = Product["channelListings"];
-export type VariantChannelListing = Variant["channelListings"];
-export type Channel = ProductChannelListing[number]["channel"];
