+++ src/products/components/ProductVariantChannels/useManageChannels.ts
@@ -1,0 +1,14 @@
+import { useState } from "react";
+
+export const useManageChannels = () => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const toggle = () => {
+    setIsOpen(p => !p);
+  };
+
+  return {
+    isOpen,
+    toggle,
+  };
+};
