--- src/products/components/ProductVariantCreatePage/useManageChannels.ts
@@ -1,14 +1,0 @@
-import { useState } from "react";
-
-export const useManageChannels = () => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const toggle = () => {
-    setIsOpen(p => !p);
-  };
-
-  return {
-    isOpen,
-    toggle,
-  };
-};
