+++ src/products/components/ProductVariantPage/ProductVariantPage.tsx
@@ -35,6 +35,8 @@ import { defineMessages, useIntl } from "react-intl";
 
 import ProductShipping from "../ProductShipping/ProductShipping";
 import ProductStocks, { ProductStockInput } from "../ProductStocks";
+import { useManageChannels } from "../ProductVariantChannels/useManageChannels";
+import { VariantChannelsDialog } from "../ProductVariantChannels/VariantChannelsDialog";
 import ProductVariantCheckoutSettings from "../ProductVariantCheckoutSettings/ProductVariantCheckoutSettings";
 import ProductVariantEndPreorderDialog from "../ProductVariantEndPreorderDialog";
 import ProductVariantMediaSelectDialog from "../ProductVariantImageSelectDialog";
@@ -157,7 +159,10 @@ const ProductVariantPage: React.FC<ProductVariantPageProps> = ({
 }) => {
   const intl = useIntl();
   const navigate = useNavigator();
-
+  const {
+    isOpen: isManageChannelsModalOpen,
+    toggle: toggleManageChannels,
+  } = useManageChannels();
   const [isModalOpened, setModalStatus] = React.useState(false);
   const toggleModal = () => setModalStatus(!isModalOpened);
 
@@ -254,7 +259,10 @@ const ProductVariantPage: React.FC<ProductVariantPageProps> = ({
                     />
                   </div>
                   <div>
-                    <VariantDetailsChannelsAvailabilityCard variant={variant} />
+                    <VariantDetailsChannelsAvailabilityCard
+                      variant={variant}
+                      onManageClick={toggleManageChannels}
+                    />
                     {nonSelectionAttributes.length > 0 && (
                       <>
                         <Attributes
@@ -401,6 +409,15 @@ const ProductVariantPage: React.FC<ProductVariantPageProps> = ({
                         handlers,
                       )
                     }
+                  />
+                )}
+                {(variant && variant.product.channelListings.length) > 0 && (
+                  <VariantChannelsDialog
+                    channelListings={variant.product.channelListings}
+                    selectedChannelListings={variant.channelListings}
+                    open={isManageChannelsModalOpen}
+                    onClose={toggleManageChannels}
+                    onConfirm={handlers.updateChannels}
                   />
                 )}
               </>
