+++ src/products/views/ProductVariant/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./ProductVariant";
+export * from "./ProductVariant";
